changed the endpoint to the correct script and saving the uploaded document to documents directory
