+++ Project/backend/codebase/monitoring/router.py
@@ -1,9 +1,12 @@
 from typing import List
 
 from fastapi import APIRouter, Depends
+from fastapi import UploadFile, File, HTTPException
 
 from monitoring.dao.healthcheck_dao import HealthCheckDAO
 from monitoring.schemas.healthcheck import HealthCheckResponse
+
+import os
 
 router = APIRouter()
 
@@ -48,3 +51,26 @@ async def get_dummy_models(
     """
 
     return await check_dao.get_all_healthchecks(limit=limit, offset=offset)
+
+
+# Endpoint for uploading PDF documents
+@router.post("/upload/")
+async def upload_pdf(file: UploadFile = File(...)):
+
+    # Check if the uploaded file type is correct
+    if file.content_type != "application/pdf":
+        raise HTTPException(status_code=400, detail="Uploaded file is not a PDF.")
+
+    # Define the directory for saving files
+    documents_directory = os.path.join(os.path.dirname(os.path.abspath(__file__)),
+                                       "documents")
+
+    if not os.path.exists(documents_directory):
+        os.makedirs(documents_directory)
+
+    # Save file
+    file_path = os.path.join(documents_directory, file.filename)
+    with open(file_path, "wb") as f:
+        f.write(file.file.read())
+
+    return {"filename": file.filename, "status": "uploaded successfully"}
